requirements.txt
# Ethical Eye Extension - Training Requirements
# Research Project: Explainable AI for Dark Pattern Detection

# Core ML/AI Libraries
torch>=1.12.0
transformers>=4.21.0
scikit-learn>=1.1.0
numpy>=1.21.0
pandas>=1.4.0

# SHAP for Explainable AI
shap>=0.41.0

# Data Balancing
imbalanced-learn>=0.9.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Word Cloud (optional)
wordcloud>=1.9.0

# Web Frameworks
flask>=2.2.0
flask-cors>=3.0.10
fastapi>=0.104.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Computer Vision + OCR
opencv-python>=4.8.0
pillow>=9.5.0
pytesseract>=0.3.10

# Data Processing
nltk>=3.7
spacy>=3.4.0

# Evaluation Metrics
scikit-plot>=0.3.7

# Progress Bars
tqdm>=4.64.0

# Configuration
pyyaml>=6.0
python-dotenv>=0.19.0

# Logging
loguru>=0.6.0

# Testing
pytest>=7.1.0
pytest-cov>=3.0.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.15.0

# Model Optimization
onnx>=1.12.0
onnxruntime>=1.12.0

# Additional Utilities
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0